ARGQGRA-468 Cannot use relative @Location when using "standalone" JUnit integration
Changed GrapheneExtension to setup ResourceProvider for URL depending on the framework integration options.
If container framework integration option is used, then ContainerCustomizableURLResourceProvider is chosen, which first delegates to URLResourceProvider of Arquillian.
If stand-alone framework inegration option is used, then CustomizableURLResourceProvider is chosen, which returns the Graphene custom URL if present.
